Add SettingsTab and SubscribersTab components for admin settings and subscriber management
- Implement SettingsTab for managing system settings, shipping zones, and payment methods.
- Add toggle switches for various settings including maintenance mode and user registration.
- Enable adding, editing, and deleting shipping zones with associated methods and cities.
- Create SubscribersTab to display and manage newsletter subscribers with sorting functionality.
- Integrate loading states and mock data for subscribers.
